Finished seed.sql file code, edited dateonly on birthdate in userInfo…

+++ models/medications.js
@@ -15,7 +15,7 @@ module.exports = function(sequelize, DataTypes){
             }
         },
         dose: {
-            type: DataTypes.FLOAT,
+            type: DataTypes.FLOAT(6,1),
             validate: {
                 notNull: true
             }
